fix(desktop): repair broken imports from partial github domain move
