+++ apps/desktop/src/features/github/components/Panel/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { GitFork as GithubIcon, Check, AlertCircle, Loader2 } from 'lucide-react';
 import { Button, Input, cn } from '@kay-am/ui';
-import { formatError } from '../errors';
-import { useAppStore } from '../store';
+import { formatError } from '../../../../errors';
+import { useAppStore } from '../../../../store';
 
 type SaveState = 'idle' | 'saving' | 'saved' | 'error';
